# openapi.json
{
  "components": {
    "schemas": {
...
      "DeletionResult": {
        "additionalProperties": false,
        "properties": {
          "audit_id": {
            "format": "uuid",
            "title": "Audit Id",
            "type": "string"
          },
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "opportunity_id": {
            "format": "uuid",
            "title": "Opportunity Id",
            "type": "string"
          }
        },
        "required": [
          "opportunity_id",
          "deleted",
          "audit_id"
        ],
        "title": "DeletionResult",
        "type": "object"
      },
...
      "FavoriteRequest": {
        "additionalProperties": false,
        "properties": {
          "favorite": {
            "title": "Favorite",
            "type": "boolean"
          }
        },
        "required": [
          "favorite"
        ],
        "title": "FavoriteRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LifecycleTargetPage": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/LifecycleTarget"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "search": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search"
          }
        },
        "required": [
          "items",
          "has_more",
          "search",
          "limit"
        ],
        "title": "LifecycleTargetPage",
        "type": "object"
      },
...
      "OpportunityDetail": {
        "additionalProperties": false,
        "properties": {
          "classification": {
            "default": "new",
            "enum": [
              "new",
              "related",
              "possible_rediscovery",
              "rediscovered",
              "updated"
            ],
            "title": "Classification",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceDetail"
            },
            "title": "Evidence",
            "type": "array"
          },
          "favorite": {
            "default": false,
            "title": "Favorite",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lifecycle_events": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/LifecycleEventDetail"
            },
            "title": "Lifecycle Events",
            "type": "array"
          },
          "lifecycle_status": {
            "default": "active",
            "enum": [
              "active",
              "merged"
            ],
            "title": "Lifecycle Status",
            "type": "string"
          },
          "merge_target_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merge Target Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "origin_pattern": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProblemPatternDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "origin_signals": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ProblemSignalDetail"
            },
            "title": "Origin Signals",
            "type": "array"
          },
          "primary_industry": {
            "title": "Primary Industry",
            "type": "string"
          },
          "provenance": {
            "additionalProperties": true,
            "title": "Provenance",
            "type": "object"
          },
          "rediscovery_target_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rediscovery Target Id"
          },
          "related": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RelatedOpportunityDetail"
            },
            "maxItems": 3,
            "title": "Related",
            "type": "array"
          },
          "report": {
            "$ref": "#/components/schemas/OpportunityAnalysis"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "scores": {
            "additionalProperties": {
              "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
              "type": "string"
            },
            "title": "Scores",
            "type": "object"
          },
          "verdict": {
            "title": "Verdict",
            "type": "string"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          },
          "version_number": {
            "title": "Version Number",
            "type": "integer"
          },
          "versions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/OpportunityVersionSummary"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "version_id",
          "version_number",
          "primary_industry",
          "revision",
          "report",
          "scores",
          "verdict",
          "evidence",
          "provenance"
        ],
        "title": "OpportunityDetail",
        "type": "object"
      },
...
      "PermanentDeleteRequest": {
        "additionalProperties": false,
        "properties": {
          "confirmation": {
            "maxLength": 43,
            "minLength": 43,
            "title": "Confirmation",
            "type": "string"
          },
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision",
          "confirmation"
        ],
        "title": "PermanentDeleteRequest",
        "type": "object"
      },
...
      "PortfolioListResponse": {
        "additionalProperties": false,
        "properties": {
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PortfolioItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "sort": {
            "$ref": "#/components/schemas/PortfolioSort"
          }
        },
        "required": [
          "items",
          "industry",
          "sort",
          "limit"
        ],
        "title": "PortfolioListResponse",
        "type": "object"
      },
      "PortfolioSort": {
        "description": "The complete supported portfolio sort surface.",
        "enum": [
          "default",
          "newest",
          "weighted_score_desc",
          "evidence_score_desc"
        ],
        "title": "PortfolioSort",
        "type": "string"
      },
...
      "ReverseMergeRequest": {
        "additionalProperties": false,
        "properties": {
          "confirmation": {
            "const": "REVERSE MERGE",
            "title": "Confirmation",
            "type": "string"
          },
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision",
          "confirmation"
        ],
        "title": "ReverseMergeRequest",
        "type": "object"
      },
...
      "StableOpportunityMetadata": {
        "additionalProperties": false,
        "description": "The only operator-editable fields on the stable opportunity row.",
        "properties": {
          "favorite": {
            "title": "Favorite",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "revision",
          "favorite",
          "updated_at"
        ],
        "title": "StableOpportunityMetadata",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Pensae Signal",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/opportunities": {
      "get": {
        "operationId": "list_opportunities_api_opportunities_get",
        "parameters": [
          {
            "in": "query",
            "name": "industry",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 240,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Industry"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PortfolioSort",
              "default": "default"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Opportunities"
      }
    },
    "/api/opportunities/{opportunity_id}": {
      "delete": {
        "operationId": "delete_opportunity_permanently_api_opportunities__opportunity_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermanentDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Opportunity Permanently"
      },
      "get": {
        "operationId": "opportunity_detail_api_opportunities__opportunity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity Detail"
      }
    },
    "/api/opportunities/{opportunity_id}/favorite": {
      "put": {
        "operationId": "set_opportunity_favorite_api_opportunities__opportunity_id__favorite_put",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableOpportunityMetadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Opportunity Favorite"
      }
    },
    "/api/opportunities/{opportunity_id}/lifecycle-targets": {
      "get": {
        "operationId": "opportunity_lifecycle_targets_api_opportunities__opportunity_id__lifecycle_targets_get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 240,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleTargetPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity Lifecycle Targets"
      }
    },
    "/api/opportunities/{opportunity_id}/merge": {
      "post": {
        "operationId": "merge_opportunity_api_opportunities__opportunity_id__merge_post",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleMutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Merge Opportunity"
      }
    },
    "/api/opportunities/{opportunity_id}/merge/reverse": {
      "post": {
        "operationId": "reverse_opportunity_merge_api_opportunities__opportunity_id__merge_reverse_post",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReverseMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleMutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reverse Opportunity Merge"
      }
    },
    "/api/opportunities/{opportunity_id}/note": {
      "put": {
        "operationId": "set_opportunity_note_api_opportunities__opportunity_id__note_put",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableOpportunityMetadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Opportunity Note"
      }
    },
    "/api/opportunities/{opportunity_id}/rediscovery-decision": {
      "post": {
        "operationId": "decide_opportunity_rediscovery_api_opportunities__opportunity_id__rediscovery_decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RediscoveryDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleMutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Opportunity Rediscovery"
      }
    },
    "/api/opportunities/{opportunity_id}/versions/{version_id}": {
      "get": {
        "operationId": "opportunity_version_detail_api_opportunities__opportunity_id__versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity Version Detail"
      }
    },
...
  }
}